Move api client test fixture to separate module
